requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.28.0
gotrue==2.12.4
httpx==0.27.2
postgrest==2.28.0
realtime==2.28.0
storage3==2.28.0
websockets==15.0.1
